Performance and accuracy update (#4)
The hclink software is now substantially faster (<2s) and lower memory (<500MB), and able to recover matches when checksums are provided instead of assigned allele codes. This is due to the use of a usearch indexed database, which does increase Docker image sizes substantially (~4/6 GB E. coli/Salmonella).

The repository has also been migrated to use `uv` to manage dependencies